move logging functions in model.py to logging.py / check 3 types of loss scales in test_loss_scale()
